requirements.txt
torch
torchvision
albumentations
opencv-python
numpy
Pillow
matplotlib
thop
